feat: add isNotDefault extension methods for core proto types (#68)
- add `isNotDefault` extension methods for core proto types

This is the hand-written portions of
https://github.com/googleapis/google-cloud-dart/pull/58. I'll follow
this up w/ generator changes (in the librarian repo), and then a PR back
to this repo that will re-generate the packages and update test and
example code (the rest of #58).

Happy to bike-shed on the extension method name(s). It's `isNotDefault`
now, but could be something like `notDefault`, ... .

cc @natebosch